services/vision_worker/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
requests>=2.31.0
# Optional (not installed by default):
# pytesseract>=0.3.10
